Added Compound. Added Unit Tests for Compound since guaranteeing behavior proves cumbersome without them. Added some Constants to Config. Added destroy()-Method to GraphicsAppCore to be able to mock GraphicApps for testing. Added Alpha Channel for Colors (needs testing). Added Width and Height at GraphicsObject Level.

+++ src/de/mi/ur/gpr/app/Canvas.java
@@ -98,7 +98,7 @@ public class Canvas extends JPanel {
 
     private void drawRectangle(Graphics2D g2d, Rectangle rect) {
         g2d.setColor(rect.getColor().asAWTColor());
-        g2d.fillRect((int) rect.getXPos(), (int) rect.getYPos(), rect.getWidth(), rect.getHeight());
+        g2d.fillRect((int) rect.getXPos(), (int) rect.getYPos(), (int) rect.getWidth(), (int) rect.getHeight());
     }
 
     private void drawCircle(Graphics2D g2d, Circle circle) {
